Added dateSet
Added Boolean for whether the date has been set. This will be the date against which expirations are compared, and it must be updated on each run of the program in order to be accurate.
I want this functionality to just pull from the system clock but I don't have time to re-learn how to do that at the moment, so for now I am leaving that to the user to input accurately.